fix: move assertOutputsMatch to BmvTestUtils.kt, shared by both test targets
The previous fix put the helper in Bmv2Runner.kt (a subprocess driver), which
was the wrong home — it coupled an infrastructure class to JUnit assertion
semantics. BmvTestUtils.kt is a dedicated shared-test-utilities file included in
both the bmv2_diff_test_suite macro output and the SaiIpInIpWcmpDiffTest target.
